Refactor single chart context exports
