+++ projects/js-packages/charts/src/components/single-chart-context/index.ts
@@ -1,9 +1,5 @@
-export {
-	SingleChartContext,
-	ChartInstanceContext,
-	useChartInstanceContext,
-	useSingleChartContext,
-} from './single-chart-context';
+export { SingleChartContext, ChartInstanceContext } from './single-chart-context';
+export { useChartInstanceContext, useSingleChartContext } from './use-single-chart-context';
 export type {
 	ChartInstanceContextValue,
 	ChartInstanceRef,
